Move legacy image pickle metrology code to serialtbx.detector.legacy_metrology
