+++ serialtbx/detector/legacy_metrology/generic_detector.py
