+++ xfel/cftbx/detector/metrology2phil.py
@@ -27,7 +27,7 @@ def metrology2phil(calib_dir, verbose):
   return sections2phil(sections, verbose)
 
 def sections2phil(sections, verbose):
-  from xfel.cftbx.detector.metrology import master_phil
+  from serialtbx.detector.legacy_metrology.metrology import master_phil
 
   # Properties of CSPad pixels (Philipp et al., 2007).  The counters
   # are 14 bits wide, and the pixels are square with a side length of
